+++ src/Triggers.php
@@ -24,7 +24,7 @@ class Triggers extends Plugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '1.0.0';
+    public string $schemaVersion = '2.0.1';
 
     /**
      * @inheritdoc